booker/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.2.3
djangorestframework==3.16.0
pydantic==2.11.7
pydantic_core==2.33.2
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.14.0
